mypage view add
