+++ src/pages/MyPage.js
@@ -1,9 +1,17 @@
 import styles from "../layout/signUp.module.css";
+import SignUpComponent from "../components/SignUpComponent";
+
 
 const MyPage = () => {
+
+ 
+  const url = "/member/update";
+
+  const updateBtn = <div><button type="submit">수정</button><button type="submit">취소</button></div>
+
     return(
         <>
-          <h2>마이페이지</h2>
+          <SignUpComponent url={url} btns={updateBtn}/>
         </>
     )
